fix(contact): Use proper event listener, remove backticks
I was still getting errors. When trying in browsers other than Chrome using the onsubmit method. After investigation it seems that the addEventListener approach has the most widespread support. After reverting to that, I found that IE was still having problems, but the source of the issue was backticks, which aren't supported in IE's javascript engine. Replacing them showed this approach worked consistently in Chrome, Firefox, Edge, and IE 11.